+++ src/index.js
@@ -4,5 +4,4 @@ export { Component } from './component';
 export { cloneElement } from './clone-element';
 export { createContext } from './create-context';
 export { toChildArray } from './diff/children';
-export { Suspense, lazy } from './suspense';
 export { default as options } from './options';
